+++ Source/ImoutoRebirth.Arachne/ImoutoRebirth.Arachne.Host/Settings/ArachneSettings.cs
@@ -6,6 +6,8 @@ public class ArachneSettings
 {
     public required DanbooruSettings DanbooruSettings { get; set; }
 
+    public required GelbooruSettings GelbooruSettings { get; set; }
+
     public required SankakuSettings SankakuSettings { get; set; }
 
     public required ExHentaiSettings ExHentaiSettings { get; set; }