+++ Source/ImoutoRebirth.Arachne/ImoutoRebirth.Arachne.Infrastructure/ServiceCollectionExtensions.cs
@@ -17,7 +17,8 @@ public static class ServiceCollectionExtensions
 {
     public static IServiceCollection AddArachneInfrastructure(
         this IServiceCollection services, 
-        DanbooruSettings danbooruSettings, 
+        DanbooruSettings danbooruSettings,
+        GelbooruSettings gelbooruSettings,
         SankakuSettings sankakuSettings,
         ExHentaiSettings exHentaiSettings)
     {
@@ -68,6 +69,7 @@ public static class ServiceCollectionExtensions
 
 
         services.AddTransient<DanbooruSettings>(x => danbooruSettings);
+        services.AddTransient<GelbooruSettings>(x => gelbooruSettings);
         services.AddTransient<SankakuSettings>(x => sankakuSettings);
         services.AddTransient<IOptions<Imouto.BooruParser.Implementations.Sankaku.SankakuSettings>>(_ =>
             Options.Create(new Imouto.BooruParser.Implementations.Sankaku.SankakuSettings